requirements.txt
torch
torchvision
matplotlib
numpy
Pillow
easyocr
